Refactor Flippable Card
